Add SOCIUS token entry
SOCIUS is a community-driven token built on the Solana blockchain to foster participation, collaboration, and growth within the Socius ecosystem.  
Designed to be a useful and accessible asset, SOCIUS enables fast and secure transactions, seamlessly integrating with decentralized wallets and platforms like Phantom and Raydium.  
This initiative aims to strengthe